+++ benchmarkIII.cpp
@@ -74,7 +74,7 @@ int main(int argc, char* argv[])
 		return -1;
 	}
 	
-	int tests = atoi(argv[1]);
+	long tests = atol(argv[1]);
 	int seed = time(NULL);
 	
 	double stdInsertTime = stdInsertBench(seed, tests);
